pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ninety-degree-rotator"
version = "0.2.0"
description = "A maubot plugin that automatically rotates images 90 degrees counter-clockwise"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "tionis", email = "tionis@example.com" },
]
keywords = ["maubot", "matrix", "bot", "image", "rotation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

dependencies = [
    "maubot>=0.4.0",
    "Pillow>=9.0.0",
    "cryptography>=3.0.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    # maubot-cli>=0.4.0,  # Install separately: pip install maubot
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/tionis/ninety-degree-rotator"
Repository = "https://github.com/tionis/ninety-degree-rotator"
Issues = "https://github.com/tionis/ninety-degree-rotator/issues"
Documentation = "https://github.com/tionis/ninety-degree-rotator/blob/main/README.md"

[tool.hatch.build.targets.wheel]
packages = ["NinetyDegreeRotator"]

[tool.hatch.version]
path = "NinetyDegreeRotator/__init__.py"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | archive
  | builds
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "maubot.*",
    "mautrix.*",
    "PIL.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"
